Home AC's - autocomplete, genre color changes, carousel changes, notification icon badge (#1)
* Search Auto complete + cached image updates
some of the homepage AC's

* carousel and genre card change

+++ lib/ui/widgets/creators/stats_box_row.dart
@@ -2,7 +2,7 @@ import 'package:d_reader_flutter/ui/widgets/creators/stats_box.dart';
 import 'package:flutter/material.dart';
 
 class StatsBoxRow extends StatelessWidget {
-  final int totalVolume;
+  final double totalVolume;
   final int issuesCount;
   const StatsBoxRow({
     Key? key,
